Add config panel navigation
