--- app/src/main/java/partners/moonshot/configpanel/ui/designsystem/joystick/JoystickActivity.kt
@@ -1,50 +1,0 @@
-package partners.moonshot.configpanel.ui.designsystem.joystick
-
-import android.media.MediaPlayer
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import dagger.hilt.android.AndroidEntryPoint
-import partners.moonshot.configpanel.R
-import partners.moonshot.configpanel.core.multimedia.SoundPlayer
-import partners.moonshot.configpanel.ui.designsystem.joystick.ui.theme.ConfigPanelTheme
-
-@AndroidEntryPoint
-class JoystickActivity : ComponentActivity() {
-
-    private lateinit var soundPlayer: SoundPlayer
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        soundPlayer = SoundPlayer(this, R.raw.excelent)
-        setContent {
-            ConfigPanelTheme {
-                JoystickScreen {
-                    soundPlayer.play()
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    ConfigPanelTheme {
-        Greeting("Android")
-    }
-}
